@@ -1,3 +1,3 @@
 export * from './src/cookies.service';
 export * from './src/cookies-options.service';
-export * from './src/cookies-options';
+export * from './src/cookie-options';
