@@ -1,18 +1,20 @@
 import { Inject, Injectable, InjectionToken, Injector } from '@angular/core';
 import { APP_BASE_HREF } from '@angular/common';
 
-import { CookiesOptions } from './cookies-options';
+import { CookieOptions } from './cookie-options';
 import { mergeOptions } from './utils';
 
 export const COOKIES_OPTIONS = new InjectionToken('COOKIES_OPTIONS');
 
 @Injectable()
 export class CookiesOptionsService {
-  private defaultOptions: CookiesOptions;
-  private _options: CookiesOptions;
+  private defaultOptions: CookieOptions;
+  private _options: CookieOptions;
 
-  constructor(@Inject(COOKIES_OPTIONS) options: CookiesOptions = {},
-              private injector: Injector) {
+  constructor(
+    @Inject(COOKIES_OPTIONS) options: CookieOptions = {},
+    private injector: Injector,
+  ) {
     this.defaultOptions = {
       path: this.injector.get(APP_BASE_HREF, '/'),
       domain: undefined,
@@ -22,7 +24,7 @@ export class CookiesOptionsService {
     this._options = mergeOptions(this.defaultOptions, options);
   }
 
-  get options(): CookiesOptions {
+  get options(): CookieOptions {
     return this._options;
   }
 }
