@@ -1,22 +1,22 @@
 import { Injectable } from '@angular/core';
 
-import { CookiesOptions } from './cookies-options';
+import { CookieOptions } from './cookie-options';
 import { CookiesOptionsService } from './cookies-options.service';
 import { safeJsonParse } from './utils';
 
 @Injectable()
 export class CookiesService {
-  protected options: CookiesOptions;
+  protected options: CookieOptions;
 
-  constructor(cookiesOptions: CookiesOptionsService) {
-    this.options = cookiesOptions.options;
+  constructor(cookieOptions: CookiesOptionsService) {
+    this.options = cookieOptions.options;
   }
 
-  put(key: string, value: string, options?: CookiesOptions): void {
+  put(key: string, value: string, options?: CookieOptions): void {
     this.cookiesWriter()(key, value, options);
   }
 
-  putObject(key: string, value: Object, options?: CookiesOptions): void {
+  putObject(key: string, value: Object, options?: CookieOptions): void {
     this.put(key, JSON.stringify(value), options);
   }
 
@@ -24,31 +24,31 @@ export class CookiesService {
     return (<any>this.cookiesReader())[key];
   }
 
-  getObject(key: string): { [key: string]: string } | string {
+  getObject(key: string): Record<string, string> | string {
     const value = this.get(key);
+
     return value ? safeJsonParse(value) : value;
   }
 
-  getAll(): { [key: string]: string } {
+  getAll(): Record<string, string> {
     return <any>this.cookiesReader();
   }
 
-  remove(key: string, options?: CookiesOptions): void {
+  remove(key: string, options?: CookieOptions): void {
     this.cookiesWriter()(key, undefined, options);
   }
 
   removeAll(): void {
     const cookies = this.getAll();
-    Object.keys(cookies).forEach(key => {
-      this.remove(key);
-    });
+
+    Object.keys(cookies).forEach(key => this.remove(key));
   }
 
-  protected cookiesReader(): { [key: string]: any } {
-    return { };
+  protected cookiesReader(): Record<string, string> {
+    return {};
   }
 
-  protected cookiesWriter(): (name: string, value: string | undefined, options?: CookiesOptions) => void {
+  protected cookiesWriter(): (name: string, value: string | undefined, options?: CookieOptions) => void {
     return () => { };
   }
 }
